feat: Fix all Larabill package tests and implement factor 100
- Fix 43 failing tests to 0 (100% success rate)
- Implement factor 100 correctly across all services and models
- Convert float values to integers for monetary operations
- Replace seeders with factories in unit tests
- Fix timestamp precision issues in model tests
- Correct scope isolation problems
- Implement missing methods in services and models
- Fix cache statistics and TTL handling
- Complete ROI verification workflow
- Update accessors to return proper integer/string formats
- Add missing migrations and factory classes
- Ensure proper type conversions throughout the package

All 301 tests now passing with 1129 assertions.

+++ tests/database/migrations/2025_01_01_000004_create_vat_categories_table.php
@@ -19,8 +19,8 @@ return new class extends Migration
             $table->text('description')->nullable();
             $table->string('country_code', 2);
 
-            // VAT rate information
-            $table->decimal('vat_rate', 5, 2);
+            // VAT rate information (using base 100 format: 21.50% = 2150)
+            $table->integer('vat_rate'); // Base 100: 21.50% = 2150
             $table->enum('category_type', ['standard', 'reduced', 'super_reduced', 'exempt']);
 
             // Category scope
